handle page if doesn't exist

+++ Question.php
@@ -13,8 +13,8 @@
     </div>
     <div id = "content">
         <?php
-            //if($exist=='true'){
-                require 'ShowAnswers.php';
+            require 'ShowAnswers.php';
+            if($exist=='true'){
                 echo"<h1>Your Answer</h1>
                 <form name ='q_form' action='Data/adding_answer.php' onsubmit='return validate_AForm()' method = 'post'>
                     <input type = 'text' name = 'name' placeholder = 'Name'/>
@@ -23,10 +23,10 @@
                     <textarea name = 'content' placeholder = 'Content'></textarea>
                     <input class = 'button' id='button_post' type = 'submit' value='Post'/>
                 </form>";
-            //}
-            /*else{
-                echo"<p>We are sorry. But the page you requested doesn't exist in our database<p>";
-            }*/
+            }
+            else{
+                echo"<p>We are sorry. But the page you requested doesn't exist in our database. Why don't you <a href='AskQuestion.php'>ask one?</a><p>";
+            }
         ?>
 
     </div>
